+++ apps/manifest-demo/3008-webpack-host/@mf-types/manifest-provider/compiled-types/apps/manifest-demo/3011-rspack-manifest-provider/src/asyncFile.d.ts
@@ -1,0 +1,2 @@
+declare const _default: "async file";
+export default _default;