+++ packages/node/src/runtimePlugin.ts
@@ -183,7 +183,7 @@ export default function () {
                 )(chunk, __non_webpack_require__, urlDirname, chunkName);
                 callback(null, chunk);
               } catch (e) {
-                callback(err, null);
+                callback(e, null);
               }
             });
             res.on('error', function (err) {
@@ -249,7 +249,7 @@ export default function () {
           }
 
           __webpack_require__.federation.runtime
-            .loadScriptNode(url, { attrs: {} })
+            .loadScriptNode(url, { attrs: { globalName: remoteName } })
             .then(function (res) {
               var federation = __webpack_require__.federation;
               var enhancedRemote = federation.instance.initRawContainer(